Add component dashboard row
